% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.lib.R
\name{load.lib}
\alias{load.lib}
\title{Loading multiple library}
\usage{
load.lib(...)
}
\arguments{
\item{...:}{package names. Quotations are unnecessary.}
}
\description{
Loading multiple library
}
\examples{
load.lib( "vegan" )
load.lib( c("vegan", "ggplot2") )

}
