% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeColPalette.R
\name{makeColPalette}
\alias{makeColPalette}
\title{Make color coresspondece table}
\usage{
makeColPalette(
  data,
  color = NULL,
  othersCol = "grey30",
  specificName = NULL,
  specificColor = "grey90",
  sortFun = sum,
  na.rm = TRUE
)
}
\arguments{
\item{data}{matrix or dataframe containing sample x abundance}

\item{color}{a vector contained color information}

\item{specificName}{You can specify the color to specific species (like 'Unidentified')}

\item{specificColor}{For the specific species color}

\item{others}{Not dominant color}
}
\description{
You can make color coresspondece table for ggplot2.
}
\examples{
data(test)
col <- palettes()
colpal <- makeColPalette(data=test[,-1], 
                            color=col, 
                            others='grey30', 
                            specificName = 'unidentified', 
                            specificColor = 'grey90',
                            sort=sum)

}
