% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveFiles.R
\name{savefile_set}
\alias{savefile_set}
\title{Setting saveFiles parameters.}
\usage{
savefile_set(
  dir = getwd(),
  type = c("rds", "csv", "txt"),
  row.names = FALSE,
  as_name = NULL
)
}
\arguments{
\item{dir}{(character): output path name.}

\item{type}{(character vector): file formats.}

\item{row.names}{(bool or character): If a character is provided, the column name is used as row names.}
}
\description{
Setting saveFiles parameters.
}
