% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subset.R
\name{Subset}
\alias{Subset}
\title{Extracting data}
\usage{
Subset(data, info, row)
}
\arguments{
\item{data}{data is matrix or data frame.}

\item{info}{info is unique infomation of specific column in data.}

\item{row}{row is single integer.}
}
\description{
This function is useful for a experiment data containing some treatment.
Similar to subset function.
}
\examples{
# If you want infomation of one iris species.
data(iris) 
unique <- unique(iris[,4:5])
a <- Subset(data=iris, info= unique, row=1)

# If you want to split infomation by each species.
a <- lapply(1:nrow(unique), function(n){Subset(data= iris, info= unique, row=n)})

}
