% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxa.mat.R
\name{Taxa.mat}
\alias{Taxa.mat}
\title{Remake the matrix to other taxonomy level.}
\usage{
Taxa.mat(
  data,
  taxa,
  taxaLabel,
  func = function(x) {
     sum(x)
 }
)
}
\arguments{
\item{data}{data is abundance table which colnames is ASV / OTU.}

\item{taxa}{taxa is taxnomy information table.}

\item{taxaLabel}{taxaLabel is a taxnomy level which you want to remake.}

\item{func}{func is function. default is sum()}
}
\description{
This function is useful for OTU/ASV data table. Similar to aggregate function.
}
\examples{
# If you want OTU table to Family level matrix.
library(microbiome) 
data(atlas1006)

data <- as.data.frame(t(otu_table(atlas1006)))
taxa <- tax_table(atlas1006)
taxaLabel <- 'Family'
a <- Taxa.mat(x,y, taxaLabel)

}
