% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{theme_md}
\alias{theme_md}
\title{ggplot2 theme}
\usage{
theme_md(
  bsize = 6,
  plssize = 0,
  lwd = 0.5,
  unitsize = NULL,
  family = NULL,
  gridcol = "grey92",
  legend.back.color = TRUE,
  expandCx = FALSE,
  expandCy = FALSE,
  expandDx = FALSE,
  expandDy = FALSE
)
}
\arguments{
\item{bsize}{: mimum size of text.}

\item{lwd}{: linewidth in element_rect.}

\item{unitsize}{: legend.key.size.}

\item{family}{: font family.}
}
\description{
theme for mardown style.
}
