% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{theme_text}
\alias{theme_text}
\title{ggplot2 theme}
\usage{
theme_text(bsize = 6, lwd = 0.5, unitsize = 1, family = NULL)
}
\arguments{
\item{bsize}{: mimum size of text.}

\item{lwd}{: linewidth in element_rect.}

\item{unitsize}{: legend.key.size.}

\item{family}{: font family.}
}
\description{
ggplot2 theme
}
\examples{

}
